Make it Py3 compatible - run stage2 of futurize

+++ oauth2client/xsrfutil.py
@@ -16,6 +16,8 @@
 
 """Helper methods for creating & verifying XSRF tokens."""
 
+from builtins import zip
+from builtins import str
 __authors__ = [
   '"Doug Coker" <dcoker@google.com>',
   '"Joe Gregorio" <jcgregorio@google.com>',
@@ -88,7 +90,7 @@ def validate_token(key, token, user_id, action_id="", current_time=None):
     return False
   try:
     decoded = base64.urlsafe_b64decode(str(token))
-    token_time = long(decoded.split(DELIMITER)[-1])
+    token_time = int(decoded.split(DELIMITER)[-1])
   except (TypeError, ValueError):
     return False
   if current_time is None:
